team-task-distribution/backend/users/views.py
from django.shortcuts import render
from rest_framework import viewsets
from .models import User, UserProfile, SocialMediaAccount
from .serializers import UserSerializer, UserProfileSerializer, SocialMediaAccountSerializer

# Create your views here.

class UserViewSet(viewsets.ModelViewSet):
    queryset = User.objects.all()
    serializer_class = UserSerializer

class UserProfileViewSet(viewsets.ModelViewSet):
    queryset = UserProfile.objects.all()
    serializer_class = UserProfileSerializer

class SocialMediaAccountViewSet(viewsets.ModelViewSet):
    queryset = SocialMediaAccount.objects.all()
    serializer_class = SocialMediaAccountSerializer
