team-task-distribution/backend/requirements.txt
Django==4.2.0
djangorestframework==3.14.0
django-cors-headers==4.3.0
psycopg2==2.9.9
python-dotenv==1.0.0
djangorestframework-simplejwt==5.3.0
celery==5.3.4
redis==5.0.1
django-redis==5.4.0
gunicorn==21.2.0
whitenoise==6.6.0 